Split progress tracker tests and add time and availability test suites
